Clicking button stores the customer details in memory
Before this, I was using the context menu to add a menu item
to the right click dropdown menu. I decided to start easier
and just have it so that when you click the copy details button
it shows buttons in the extension box that you can press 'copy' for
and copies it to clipboard.

This commit takes the customer details from the rfq page and
stores them in memory.